handle saving the privacy policy page setting
since the privacy policy page setting does not actually belong to this page's options group -- and, in fact, there is no options group, because that page is weird -- we need to intercept the submitted value and update the option manually. This is actually more or less what the hard-coded page in core does.